requirements.txt
fastapi
uvicorn
redis
httpx
pytest
pytest-asyncio
nats-py
